requirements.txt
torch.version==2.4.1+cu121
sklearn.version==1.3.2
pyyaml==6.0.2
timm==1.0.12
numpy==1.24.4
matplotlib==3.7.5
pandas==2.0.3
scipy==1.10.1

